+++ mayan/apps/templating/apps.py
@@ -1,10 +1,12 @@
 from django.apps import apps
+from django.utils import timezone
 from django.utils.translation import gettext_lazy as _
 
 from mayan.apps.acls.classes import ModelPermission
 from mayan.apps.common.apps import MayanAppConfig
 from mayan.apps.common.menus import menu_list_facet
 
+from .classes import TemplateContextEntry
 from .links import link_document_template_sandbox
 from .permissions import permission_template_sandbox
 
@@ -30,4 +32,10 @@ class TemplatingApp(MayanAppConfig):
 
         menu_list_facet.bind_links(
             links=(link_document_template_sandbox,), sources=(Document,)
+        )
+
+        TemplateContextEntry(
+            always_available=True,
+            description=_(message='Current date and time'), name='now',
+            value=timezone.now
         )