+++ mayan/apps/user_management/apps.py
@@ -16,9 +16,10 @@ from mayan.apps.common.menus import (
 from mayan.apps.dashboards.dashboards import dashboard_administrator
 from mayan.apps.events.classes import EventModelRegistry, ModelEventType
 from mayan.apps.logging.classes import ErrorLog
-from mayan.apps.metadata.classes import MetadataLookup
+# ~ from mayan.apps.metadata.classes import MetadataLookup
 from mayan.apps.navigation.classes import SourceColumn
 from mayan.apps.rest_api.fields import DynamicSerializerField
+from mayan.apps.templating.classes import TemplateContextEntry
 from mayan.apps.views.column_widgets import TwoStateWidget
 
 from .dashboard_widgets import (
@@ -169,11 +170,11 @@ class UserManagementApp(MayanAppConfig):
         EventModelRegistry.register(model=Group)
         EventModelRegistry.register(model=User)
 
-        MetadataLookup(
+        TemplateContextEntry(
             description=_(message='All the groups.'), name='groups',
             value=get_groups
         )
-        MetadataLookup(
+        TemplateContextEntry(
             description=_(message='All the users.'), name='users',
             value=get_users
         )